perf(landing): populate GitHub contribution tile titles on client
Since the dates are redundant, populating them in the HTML sent from the
server is wasteful. We can calculate the dates on the client by simply
sending the first date and extrapolating from there. This lets us remove
the title attribute from the HTML and still populate it on the client,
saving ~50 kB.